+++ marketplace/core/types/channels/generic/views.py
@@ -1,10 +1,8 @@
 """  this view manages all available channels configuring them in a generic way """
-
 from rest_framework.response import Response
 from rest_framework.decorators import action
 from rest_framework import serializers
 from rest_framework import viewsets
-from rest_framework.exceptions import APIException
 
 from .serializers import GenericChannelSerializer, GenericConfigureSerializer
 
@@ -14,6 +12,10 @@ from marketplace.flows.client import FlowsClient
 from marketplace.applications.models import AppTypeAsset
 
 from . import type as type_
+
+from django.conf import settings
+
+IMPORTANCE_CHANNELS_ORDER = settings.IMPORTANCE_CHANNELS_ORDER
 
 
 class GenericChannelViewSet(views.BaseAppTypeViewSet):
@@ -133,17 +135,9 @@ class GenericAppTypes(viewsets.ViewSet):
     """
 
     def list(self, request):
-        try:
-            client = FlowsClient()
-            response = client.list_channel_types(channel_code=None)
-            response.raise_for_status()
-
-        except Exception as _error:
-            error_message = str(_error)
-            status_code = getattr(_error, "response", {}).status_code or 500
-            raise APIException(error_message, code=status_code) from _error
-
-        channel_types = response.json().get("channel_types")
+        client = FlowsClient()
+        response = client.list_channel_types(channel_code=None)
+        channel_types = sort_channel_types(response.json().get("channel_types"))
         return Response(channel_types, status=response.status_code)
 
 
@@ -169,3 +163,23 @@ def search_icon(code):
             icon_url = None
 
     return icon_url
+
+
+def sort_channel_types(channel_types):
+    """Receives a dictionary of channels and returns
+    them ordered by importance and alphabetical order of the channel code"""
+
+    importance_order = IMPORTANCE_CHANNELS_ORDER
+    keys_sorted = sorted(set(importance_order + list(channel_types.keys())))
+    items = [(key, channel_types[key]) for key in keys_sorted]
+
+    # Sort the list of tuples in order of importance and by name within the attributes
+    items_sorted = sorted(
+        items,
+        key=lambda x: (
+            importance_order.index(x[0]) if x[0] in importance_order else float("inf"),
+            x[0],
+        ),
+    )
+    channel_types_sorted = {key: value for key, value in items_sorted}
+    return channel_types_sorted